Split open view workflows into focused mixins.
